refact: building main page
