+++ src/components/organisms/CardList/_cardList.sass
@@ -1,8 +1,8 @@
 @use '../../../styles' as *
 
-.main
+.cardList
 
-  &__cardHeader
+  &__header
     display: flex
     justify-content: space-between
     font-family: $ff-base
@@ -23,7 +23,7 @@
     &__selector
       align-self: center
 
-  &__cardList
+  &__list
     @include center-center
     padding: 0
     list-style-type: none
