+++ src/components/organisms/Footer/Footer.tsx
@@ -1,0 +1,21 @@
+import './_footer.sass'
+
+const Footer = () => {
+  return (
+    <footer>
+      <div>
+        <p>Terms and Conditions</p>
+        <p>Privacy Policy</p>
+        <p>Contact Us</p>
+      </div>
+
+      <div>
+        <p>Follow us</p>
+        <img src="" alt="Facebook" />
+        <img src="" alt="Twitter" />
+      </div>
+    </footer>
+  )
+}
+
+export { Footer }
