+++ src/components/organisms/Footer/_footer.sass
@@ -1,0 +1,1 @@
+@use '../../../styles' as *
