+++ src/components/organisms/index.ts
@@ -1,4 +1,5 @@
 export { Card } from './Card/Card'
-export { Main } from './Main/Main'
+export { CardList } from './CardList/CardList'
 export { InfoBanner } from './InfoBanner/InfoBanner'
 export { FeedbackForm } from './FeedbackForm/FeedbackForm'
+export { Footer } from './Footer/Footer'
