+++ src/components/pages/Main/Main.tsx
@@ -1,0 +1,19 @@
+import './_main.sass'
+
+import { InfoBanner, CardList , FeedbackForm, Footer } from '../../organisms'
+
+const Main = () => {
+  return (
+    <>
+      <InfoBanner />
+
+      <CardList />
+
+      <FeedbackForm />
+
+      <Footer/>
+    </>
+  )
+}
+
+export { Main }
